Fix contributor page loading screen styling
Before, when you loaded the contributor page, you'd see a loading
screen where there was a component above the loading page which
looked awkward. This commit fixes that styling bug.

I was able to figure out how to fix the styling by opening
up Chrome devtools, going to the Elements tab, placing a break
point on subtree modifications on the body element, then I refreshed
the page and was able to add styles. It turns out that adding a simple
z-index: 1 was all it took to fix this.