Add resolver for updating user names
This commit adds an Apollo resolver for updating the user's name and username in the database using Prisma.

To implement this functionality, we added a new `updateUserNames` mutation to the GraphQL schema. This mutation accepts arguments for the user's name and username and uses Prisma to update the corresponding fields in the database.

We also added tests for the `updateUserNames` mutation to ensure that it is functioning correctly.

This change allows users to update their name and username through the application, improving the usability and functionality of the app.